Fix windows warning. (#332)

+++ examples/dom/canvas.cpp
@@ -30,7 +30,7 @@ int main(int argc, const char* argv[]) {
   // Plot a function:
   std::vector<int> ys(100);
   for (int x = 0; x < 100; x++)
-    ys[x] = 80 + 20 * cos(x * 0.2);
+    ys[x] = int(80 + 20 * cos(x * 0.2));
   for (int x = 0; x < 99; x++)
     c.DrawPointLine(x, ys[x], x + 1, ys[x + 1], Color::Red);
